housekeeping: refactor body index form screen into separate files
